# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse:
    properties:
      error:
        type: string
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.ValidationError'
        type: array
    type: object
...
  github_com_Aliizi83_vohu_internal_platform_shared.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 40901
    - 50001
    type: integer
    x-enum-varnames:
    - ResultSuccess
    - ResultValidationError
    - ResultAuthError
    - ResultForbiddenError
    - ResultNotFoundError
    - ResultConflictError
    - ResultInternalError
  github_com_Aliizi83_vohu_internal_platform_shared.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
  internal_platform_providerkey.Provider:
    enum:
    - gemini
    - anthropic
    - openai
    type: string
    x-enum-varnames:
    - ProviderGemini
    - ProviderAnthropic
    - ProviderOpenAI
  internal_platform_providerkey.Response:
    properties:
      baseUrl:
        type: string
      provider:
        $ref: '#/definitions/internal_platform_providerkey.Provider'
      workspaceId:
        type: string
    type: object
  internal_platform_providerkey.SetKeyRequest:
    properties:
      apiKey:
        type: string
      baseUrl:
        type: string
      provider:
        allOf:
        - $ref: '#/definitions/internal_platform_providerkey.Provider'
        enum:
        - gemini
        - anthropic
        - openai
      workspaceId:
        type: string
    required:
    - apiKey
    - provider
    type: object
...
info:
  contact: {}
  description: HTTP API for Vohu's multi-user platform — users, roles, polymorphic
    resource access, SSH connections, conversations/chat, provider keys, custom models,
    and the agent tool catalog. Every endpoint responds through shared.BaseResponse
    (see internal/platform/shared/response.go) — {result, success, resultCode, validationErrors,
    error} — regardless of success or failure.
  title: Vohu Platform API
  version: "1.0"
paths:
...
  /provider-keys:
    get:
      description: Lists every account-wide default provider key (never the key value
        itself). Requires wildcard "manage" access on resource type "provider_key".
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/internal_platform_providerkey.Response'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: List global provider keys
      tags:
      - provider-keys
    post:
      consumes:
      - application/json
      description: Sets (or replaces) the account-wide default API key for a provider
        — used for any user who hasn't set their own personal key. The key is encrypted
        at rest and never returned by any response. Requires wildcard "manage" access
        on resource type "provider_key".
      parameters:
      - description: Provider key
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_platform_providerkey.SetKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Set the global provider key
      tags:
      - provider-keys
  /provider-keys/{provider}:
    delete:
      description: Removes the account-wide default key for a provider. Chats using
        it fall back to any other configured key. Requires wildcard "manage" access
        on resource type "provider_key".
      parameters:
      - description: Provider
        enum:
        - gemini
        - anthropic
        - openai
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Remove the global provider key
      tags:
      - provider-keys
  /provider-keys/me:
    get:
      description: Lists the caller's own personal provider keys (never the key values
        themselves).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/internal_platform_providerkey.Response'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: List my provider keys
      tags:
      - provider-keys
    post:
      consumes:
      - application/json
      description: Sets (or replaces) the caller's own personal API key for a provider
        — always takes priority over the global default for that caller's own chats.
        The key is encrypted at rest and never returned by any response.
      parameters:
      - description: Provider key
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_platform_providerkey.SetKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Set my provider key
      tags:
      - provider-keys
  /provider-keys/me/{provider}:
    delete:
      description: Removes the caller's own personal key for a provider. Their chats
        fall back to the global default, if any.
      parameters:
      - description: Provider
        enum:
        - gemini
        - anthropic
        - openai
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Remove my provider key
      tags:
      - provider-keys
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token from POST
      /auth/login or /auth/refresh.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"